refactor: move analytics mutations into their own file
